feat: Parallel transform step accepts optional initializer function (#23)
This adds an optional `initializer` argument to the parallel transform step
which allows a custom initialization function to be run when a process is
spawned. The custom initialize function can also be provided via the
ConsumerStrategyFactory.

The primary motivation for this is to allow Snuba to pass a function that
sets up the Sentry SDK immediately when a process is spawned so that
any logs or errors can be reported to Sentry.