feat: Keep track of timestamps together with offsets in a batch
This is one of the least invasive ways I've found so far to make the
timestamp associated with an offset availble to the collector processing step.
The end goal of this is to allow Snuba to produce that timestamp to the commit log
during the commit offsets phase, so that it can later be used for subscription scheduling.

We already keep track of the hi and lo offset for each batch and use that to commit offsets.
With this change we will also keep track of the associated timestamp together with it's
corresponding offset number.