feat(metrics) Add a gauge set to the number of running processes (#33)
This is meant to have a metric that provides the number of processes to be used to normalize other metrics.
Example:
We have the number of processes being actively in use at a given point in time. This can be used both as a utilization and as a saturation metric if we had the number of available process as well. So we can have alert on the amount of processes in use vs the total.