feat: Support incremental assignments in stream processor (#58)
With the incremental assign / revoke strategy the assignment callback can
be triggered without a corresponding revocation beforehand. In that case
the processing strategy will still be active. When this happens, we close
the strategy before recreating it to make the behavior compatible with
the eager strategy. On revocation, we recreate the strategy after closing
if the consumer still has another partition assigned.