fix(dlq): Policy closure added + Producer closed (#68)
The produce policy takes in a producer which is never closed, this leaves the consumer hanging on shutdown. The producer is not recreated upon consumer rebalancing as well.

Added self.__producer.close() to the produce policy join method, this fixes the hanging issue. Introduced the policy closure to the Consumer Strategy Factory in order to recreate the policy entirely every time the create() method is called (eg when rebalancing).

Introduced close() and terminate() methods to Policies to be used when DLQ is closed/terminated.