feat: Support throttle option in commit function
Currently most consumers that use this library need to manually throttle
commits, as it's important to commit offsets to Kafka only periodically
and not on, say, every message.

The goal of this change is to make it easier to write consumers that do
this. The commit callback now supports an additional `throttle_sec` parameter.
If passed, offsets will be staged but committing won't be done unless that
duration has elapsed since the last commit.