feat: Remove the synchronized consumer
The synchronized consumer should not be part of Arroyo. It was previously
used only by Snuba's legacy subscriptions infrastructure. This is now
rebuilt using regular (non-synchronized) consumers for performance reasons.
Because of those performance reasons, this is not really a pattern we want to
encourage by providing it out of the box in this library.