feat: Introduce commit policies (#106)
This is an attempt to simplify throttling commits in consumers.
Generally speaking it is not safe for high throughput consumers to
commit every offset. Today in order to achieve this with Arroyo, every
consumer needs to manually implement throttling by themselves. This is verbose
and easy to get wrong. This change aims to make it easier to do the right
thing.

With this change, the stream processor takes an optional commit policy
argument. The commit policy can specifies the minimum time or minimum messages
between commits.

In order to avoid throttling during shutdown, the commit function also now
supports a force flag to override the commit policy / throttling behavior.
This should be used during consumer shutdown sequence as we no longer want
to throttle in that scenario and just want to commit the latest offsets.

This change is designed to be backward compatible.
The default behavior if no commit policy is passed remains the same as
it was previously (no throttling, every offset committed).