ref: Move strategies out of streaming module (#114)
The streaming module is kind of redundant and just makes imports more verbose - this whole library is about streaming anyway and the batching strategies will be getting deprecated. This structure added confusion for users about when to use batching vs streaming. All the strategies will now live directly under the strategies module and are generally safe to combine together.