feat: Make commit policy a required argument (#116)
StreamProcessor will now require commit policy to be passed as a required argument. Previously the default (IMMEDIATE) was applied, however this is not a good default as it may lead to way too much committing. The user is now forced to pick a policy.

Also pulls out the docs on committing offsets into their own section, as it is easier to find than buried inside getstarted.