feat: Split the Message interface to better support batching steps
This is an attempt to address the problems described in https://github.com/getsentry/arroyo/issues/125.
It would be a major breaking change of Arroyo's API.

Today the `Message` interface is tied to a specific broker message (with a single
partition, offset and timestamp). This doesn't make a lot of sense for strategies
where the Message does not actually represent a single message but a batch of messages
that have been grouped together - or, really, any scenario where a message no longer
maps cleanly to a single underlying message from the Kafka broker.

This solution involves splitting the `Message` interface apart into `Message`
and `BrokerPayload` to better serve the two distinct purposes. Now `Message`
only has a `committable` property which is a mapping of partitions and offsets.
This committable allows the message to easily span partitions. Since
committing offsets is generally always the last step of a processing pipeline
after all processing steps are done, this remains as part of the Message interface -
it needs to always be passed through to every strategy so we know which offsets
to commit at the very end after we are done with all the processing steps.

However messages received from the broker actually do map to a singular partition,
offset and timestamp. The separate `BrokerPayload` class is introduced to address this.
The first strategy in the pipeline will get a `Message[BrokerPayload[TPayload]]`;
but thereafter the user can just pass a `Message[TPayload]` so the user can choose
to include or exclude the BrokerPayload fields in later steps as needed.

I think that the main downside of this approach is that we are now storing additional
information in every message about the partition and offset even when we are only dealing
with singular messages (the committable is separate to the message's partition and offset
data).