feat: Avoid storing entire messages in the produce step (#132)
We were storing entire messages in the queue of the produce step just so we know what offset to commit once the future is done. Since the messages can be quite large and the queue quite long this seems wasteful. Now we are just storing the parts we actually need to commit.