feat: Add RunTask strategy
This is the most basic strategy that Arroyo provides. The rationale
behind this addition is to make the library even simpler and reduce
the amount of boilerplace for simple use cases.

Specifically, we noticed that many people are using the
BatchProcessingStrategy and AbstractBatchWorker combination almost
exclusively for non batching use cases, purely because it allowed for
using Arroyo without requiring a custom processing strategy.
This is supposed to provide an easy replacement in those scenarios.

This implementation is almost identical to the previous transform step,
however this way explicitly allows for side effects. Now TransformStep
is simply an alias for RunTask which is kept for backward compatibility
and compatibility with the ParallelTransformStep.