feat: Collect / ParallelCollect supports a next step
Instead of committing offsets after collect is done, now a next step can
be passed to collect. This enables the collect step to be more flexible
as it no longer needs to be the last step in a chain of strategies.

The other immediate motivation for this is to support a refactor of Snuba's
consumer with commit log, which needs to do something else (produce offsets)
after the collect step (which writes to ClickHouse) but before committing.

Strategies built using the ConsumerStrategyFactory/KafkaConsumerStrategyFactory
(which will soon be deprecated anyway) are backward compatible and will work in
the same way (commit immediately after collect).