feat: Add RunTask strategy (#147)
This is the most basic strategy that Arroyo provides. The rationale behind this addition is to make the library even simpler and reduce the amount of boilerplate for simple use cases.

Specifically, I've noticed that many people are using the BatchProcessingStrategy and AbstractBatchWorker combination almost exclusively for non batching use cases, purely because it allowed for using Arroyo without requiring a custom processing strategy. This is supposed to provide an easy replacement in those scenarios that doesn't have the downside of
unnecessarily waiting for batches to get filled.

This implementation is almost identical to the previous transform step, however RunTask explicitly allows for side effects. Now TransformStep is simply an alias for RunTask which is kept for backward compatibility and for symmetry with the ParallelTransformStep.