feat(Collect) Attempt to redesign the Collector step (#127)
The current CollectStep is fairly confusing. It is not only batching messages but it embeds the logic that executes application code when the batch is full. This logic is itself represented as a ProcessingStrategy, but the Collector is not connected to this logic through ProcessingStrategy chaining as usual.

The idea here is to make the collect idea more modular.
It introduces a BatchStep and an UnbatchStep. These processing strategies are simpler. The first accumulates messages in a batch and submit the batch to the next step, the second does the opposite. It takes a batch and submits individual messages.