fix(commit_policy): Calculate elapsed timerange correctly
The first argument to should_commit is a duration/timedelta, not an
absolute timestamp.

This might explain why we are seeing severe performance regression when
using commit policy.

Some tests are failing, but i wanted to open this PR early on to get
some validation that i'm not totally confused?