ref: Remove ``Position`` and stop passing timestamps to the commit function
Committing offsets never requires the original message's timestamp. All you should
need is the offset for the partition. The only reason this interface was ever
needed was to support Snuba's KafkaConsumerWithCommitLog, which implements part of
it's logic in a consumer commit function (where it produces to the commit log topic)
instead of putting that logic into the main processing strategy like regular Arroyo
consumers do.

Now we are in the process of removing that from Snuba (https://github.com/getsentry/snuba/pull/3511
and https://github.com/getsentry/snuba/pull/3507) so we can remove `Position` entirely
from Arroyo now.

In addition to making for a cleaner interface, we also expect some small performance
benefits as every call to stage offsets previously was making copies of the whole position
object, even though only the offset part was really needed.