feat: Add Reduce strategy (#157)
The Reduce strategy aims to provide a more generalized version of the BatchStep, which allows the user to return any generic accumulated value not just a ValuesBatch.

It is designed to be pretty similar to how reduce works in many programming lanugages - the user provides an initial value and an accumulator function. Reduce will periodically close batches of accumulated values and submit those batches to the next step.

This allows Reduce to be used anywhere that Collect and ParallelCollect is used today. We should consider fully deprecating Collect in favor of Reduce, since now Collect doesn't offer anything more than a Reduce chained to a RunTask, and ParallelCollect can be nothing more than a Reduce chained to a RunTaskInThreads.

I've left the BatchStep/UnbatchStep interface unchanged, but updated the BatchStep implementation to use Reduce under the hood.