feat: Remove support for cooperative-sticky partition assignment strategy
We experiemnted with supporting the cooperative-sticky partition assignment strategy
but it didn't actually work well with Arroyo's design. Specifically, if a partition
was not undergoing reassignment, Arroyo would attempt to continue normal processing,
including committing offsets, on that partition. But this is not allowed due to the
"group-wide synchronization barrier".

We'll probably want to wait for https://cwiki.apache.org/confluence/display/KAFKA/KIP-848%3A+The+Next+Generation+of+the+Consumer+Rebalance+Protocol
to be implemented in Kafka before trying this again.