feat(reduce): Support greater flexibility in initial value
The motivation for this change is to make it easier to provide a custom
initial value for the Reduce step. The current implementation uses deepcopy
however some complex objects cannot be easily deepcopied, or may not need to be
completely deep copied on each batch.

The new implementation is much closer to how the CollectStep works today -
CollectStep takes a `step_factory` callable which is called at the start
of each batch.

This change will allow us to more easily drop in Reduce where CollectStep
is used today, and fully deprecate CollectStep and ParallelCollectStep.