feat: Configurable strategy join timeout (#187)
This change allows the user to configure the join timeout of a consumer strategy.

Currently, the timeout is hardcoded to None. The implication of this is that
consumers can take a long time to finish rebalancing. This time may stack up,
especially during deployments where multiple rebalancings will occur back to back
as consumers are gradually rolled out and shut down. In some scenarios, it may be better
to abandon in-flight messages and just immediately shut down.

This change keeps same default as before, but now consumers can pass a custom strategy
join timeout if they want recovation callbacks to finish faster.