fix(filter-strategy): Options to commit filtered messages, take 2 (#185)
Alternative to #177, reimplements the same exact thing without touching Message. Instead, every strategy has to declare whether it can support filtered messages explicitly.

Strategies still have to explicitly handle filtered messages. There's no way around that IMO: If there were some wrapper logic like @fpacifici proposed in the other PR, I imagine it would be hard to implement that in a way that does not potentially forward messages out of order. If messages are forwarded out of order, this has implications for committable.

However, I think this PR is a more solid basis for implementing a "wrapper strategy" (or even a class decorator) if we can think of a way to do that. In the other PR, every strategy inherently has to support filtered messages. In this PR, a strategy can opt to not support it, but right now any composition of it with other strategies is unlikely to typecheck.

I personally like this PR more than the other one.

Either PR is a breaking change. In both cases, strategies in Snuba and Sentry need to be updated to typecheck again. There is no breakage at runtime though unless the new filter options are enabled.