feat: Add ability to assign consumer to a subset of partitions
Today Arroyo's consumer only works with auto assignment.

This is closer to how the on assignment callback works in the actual
Kafka consumer API.

The immediate motivation for this change is to support a dead letter queue
consumer implementation which will seek to specific partitions/offsets and
fetch the invalid messages from those positions.