feat: Add unfold strategy
This strategy is the opposite of Reduce. The objective is to provide a more flexible
way to extract a number of messages from a single message than UnbatchStep
currently allows. With UnbatchStep, the input message must be `ValuesBatch`,
which is an array of BaseValue from the BatchStep. Unfold
allows for any user defined type and user provided generation function.

This change is motivated by a recent attempt to use Batch/Unbatch in the
Sentry indexer (https://github.com/getsentry/sentry/pull/44721). The other
strategies in the pipeline expected a slightly different format to what the
BatchStep outputs. Switching out single steps in this pipeline without changing
the rest would be easier if we were able to define both custom fold and unfold
functions.