feat(dlq): A DLQ implementation that keeps a copy of raw messages in a buffer
In this implementation, the stream processor holds onto a copy of all the
raw messages until they are determined to be valid (when commit is called
by the strategies).

Downsides are that an extra copy of messages is being kept, and the commit
synchronization code is a bit complex. Taking suggestions on potential
ways to avoid or improve here.