ref: Make stream processor / producer / consumer not generic
Throwing this up as an alternative idea to https://github.com/getsentry/arroyo/pull/200.

I don't think there's any reason why stream processor, producer and
consumer actually need to be generic. It makes typing more annoying and
AFAIK there's no reason why we shouldn't enforce that a message consumed
or produced needs to be a KafkaPayload - that is - a tuple of key,
value (bytes) and headers.