ref: Split run task, update strategies docs page
The run_task file is getting unwieldy, and is about to become longer
with changes needed for the DLQ. Let's split it up so reading code is
more manageable and the processing strategy docs page is better structured.

There are no functional code changes here, just moving the classes to separate
modules before working on them.