ref: Remove the FileMessageStorage backend
This was only ever intended for use in tests but we never actually use
it in any tests since we always use the in memory storage instead.

The way offsets work in this backend istoo different to the Kafka and
Local/memory implementations, so I suspect many tests would even pass
if we attempted to drop it this backend instead.