feat(dlq): A DLQ implementation that keeps a copy of raw messages in a buffer (#202)
In this DLQ implementation, the stream processor holds onto a copy of all the raw messages until they are determined to be valid (when commit is called by the strategies).

The `RunTask` and `KafkaMessageDecoder` processing strategies have been updated to raise InvalidMessage exception if they encounter an invalid message. They both also pass a FilteredPayload with the offset of the DLQ'ed message through the pipeline on the subsequent call to poll so that DLQ'ed messages get committed in the correct order. Support for other strategies such as `RunTaskInThreads` and `RunTaskWithMultiprocessing` will be added as a follow up.

The KafkaDlqProducer implementation has also been updated to include additional metadata in the headers of the message produced to the DLQ. The "original_partition" and "original_offset" headers contains information about the original partition and offset the invalid message came from.

If no DLQ policy is passed, messages are never buffered. This is because there is a small overhead to buffering messages and we prefer to skip it if it's not needed.