fix(dlq): Fix crash during join when InvalidMessage in flight
During the DLQ test, we hit the `assert not self.__closed` error during
rebalancing as we were attempting to submit filtered payload messages
to strategies that we already closed.

This change moves the next_step.submit(filtered_payload) call before
the next_step.close() call to avoid this bug.

If any strategy raises InvalidMessage during join, the guard will drop it -
this should be safe to do as it will be reprocessed after the partition is
reassigned and the invalid message should only get produced and committed once.