feat: Count how many times consumer spins
See SNS-2258

As part of resolving INC-344, we are still trying to figure out whether
the consumer was spinning and hogging the GIL of the main thread, and
whether that was responsible for the consumer grinding to a halt and
therefore the root cause of the incident.

Let's just put a metric in to see how often arroyo spins. While we are
"pretty sure" that the consumer spins, we are not sure about it being
the root cause, and IMO confidence as to whether it even happened can be
higher. Afaik the only indicator we have is elevated CPU usage (could
have been caused by something else) and increased pause time.