fix: Do not log error when consumer is already crashing
When the consumer is crashing, we are terminating the strategy in
`StreamProcessor.run`. Later, rdkafka does partition revocation. At that
point the strategy is already None.

I currently believe this behavior has been there since a very long time
and does not cause any practical issues in production we are currently
seeing (such as consumers stalling and not processing any messages), and
rather that we are seeing this as a side-effect of exception logging in
rdkafka callbacks having been improved recently.

However, we are seeing this exception now in Sentry, and it should be
fixed.