perf(dlq): Improve performance when there are many invalid messages
Previously processing invalid messages was very slow since each message
was produced one a time and blocks until it is done. This change allows
the producer to do some batching of messages to be produced, and only
waits for messages to be produced prior to committing, or if there are
too many pending messages enqueued.