feat(metrics): Allow reconfiguring the metrics backend
There are some rare situations where we might want to reconfigure
the metrics backend to something else. This came up in the DLQ
consumer code, which can be instructed to consume from multiple queues
and apply different policies. In this scenario, we want to tag the
metrics very differently each time.