feat(metrics): Allow reconfiguring the metrics backend (#249)
There are some rare situations where we might want to reconfigure the metrics backend to something else after a while. This came up in the DLQ consumer code, which can be instructed to consume from multiple queues and apply different policies. In this scenario, we want to tag the metrics very differently each time.