feat: Random idea to avoid recreating strategies all the time
Based on comment from @untitaker in https://github.com/getsentry/arroyo/pull/254#issuecomment-1593185164.
Throwing this up as draft as it's not yet a properly fleshed out idea.

Motivation is that we need the on_revoke and on_assign callback to
be as fast as possible, and most of the time the strategies can be
reused as-is.