fix(reduce): Add missing call to next_step.terminate()
When a reduce strategy is terminated, the next strategy should also be
terminated.

We think that in INC-443, this caused the RunTaskInThreads in Snuba to
not be terminated at all, leaving us with background threads that
inhibited the shutdown of the process.

Also add more generic tests around close+join, and align arguments in
all strategies to be by-keyword so that the assertion passes.