feat: Add global shutdown handler for strategy factory
In sentry's unified consumers, there is no place to register a global
shutdown handler. Add this to arroyo directly in order to standardize
it, though technically there's no use for such a shutdown outside of
unified consumers since in other scenarios, the user manages the
lifecycle of the application themselves.