Make producer thread a deamon
Sometimes a program may have to shut down before the strategies are created. This producer thread starts running as soon as the producer is instantiated even before it is added to a strategy.  To simplify cleanup, make it a daemon thread so that it exists when the parent process exists and doesn't block exiting.