fix: Ensure carried over message is in buffer
Since `_run_once` can early return (https://github.com/getsentry/arroyo/blob/6286c7921978065edeb5ce72d829031f952d2e5e/arroyo/processing/processor.py#L369)
it was possible that a message was never placed in `self.buffered_messages`. If we try to retreive it later, it can crash
the consumer.

This is suspected to be the cause of the `Invalid message not found in buffer` messages we saw in prod.