fix(dlq): Gracefully handle case of no valid messages
In the case where all messages on the kafka pipeline are bad, it is possible
that there is no valid message. Today, the code would crash because of division
by zero error. Instead, we should reject the messages and let the consumers
backlog because zero valid messages indicates systemic problems in the
pipeline.