fix: Revert change when consumer is paused
This reverts part of the change made in https://github.com/getsentry/arroyo/pull/296,
which removed the check for the carried over message, when validating that
the consumer is paused. This change was inadvertently introduced and is not actually
needed to achieve th goal of throttling calls to consumer pause/resume.

This caused the consumer to crash with InvalidStateError so let's revert this part.