fix: Default input block size should be larger
We were under the assumption that we can pick an arbitrarily low value
here since it will be automatically resized. However, in the case where
a single message is larger than the block size, the consumer can still
crash.

This is too hard to fix right now and not worth it, so hardcode it to
16M (which is more than the max Kafka message size), and hope it never
happens again.

The original purpose of dynamically resizing input blocks was to make
performance tuning simpler, and IMO we still have achieved that.