fix(dlq): Actually respect the DLQ limits
While working on porting this code to the Rust consumer, I noticed
that the DLQ limits are not working in Python as expected. This
is an attempt to fix that. Specifically, limits are being ignored.
Even if one is defined, the Arroyo processor behaves as if there was
no limit set and simply forwards all invalid messages to the DLQ.