fix(filter): Handle MessageRejected correctly
When joining a FilterStep, it can happen that the upstream strategy
raises MessageRejected. We need to deal with that somehow, and in this
PR I'm just suppressing that exception.

Also fix a severe bug where MessageRejected was propagated through
submit(), but the MessageRejected was for the wrong message, causing
double-submission of messages.