fix(multiprocessing): Reset pool if tasks are not completed
If multiprocessing tasks are not completed within the timeout specified,
we need to reset the pool to avoid state being carried over between
assignments.

This change also defers the initial creation of the multiprocessing pool
until the first message is received, which moves it out of the assignment
callback (which needs to be fast).