fix(rust): Suppress error when manually resetting offsets (#5186)
* fix(rust): Suppress error when manually resetting offsets

This fixes the same bug as https://github.com/getsentry/snuba/pull/5181,
but with a test and without the potential to suppress all kinds of
poll() errors mid-operation. Still a gross hack though.

Also fix a bug where consumer.tell() would be out of date once
consumer.commit_offsets was used. We don't really use tell() in prod so
it didn't come up.

* fmt

* tighter errorhandling

* fmt