fix(multiprocessing): Increase default block sizes
There are some Sentry topics for which we allow a single message to be as
large as 25 MB. In cases like these, using the autoresizing of block sizes
could fail if the default block sizes are smaller than the maximum size of
a single message. This change modifies the default block sizes to be atleast
as high as the maximum size of a single message.