ref: Use coarsetime consistently (#366)
* ref: Use coarsetime consistently

coarsetime is a crate that allows you to fetch the current time with an
in-memory cache. This tends to be faster than actually getting the
current time.

Previously we were using Instant::now_without_update, which does not
actually give us any perf benefits. Let's use ::recent everywhere and
update the timestamp in the StreamProcessor.

* fix tests and examples

* fix tests