fix(unfold): Do not forward empty committable
In https://github.com/getsentry/arroyo/pull/371 we changed the API for
Unfold to allow for custom committable. However, we also made it
mandatory to figure out a committable at all in order for committing to
work.

The callbacks for unfold do not get access to the original committable,
so it's actually not possible to satisfy the 99% usecase with this.

cc @mj0nez