ref(produce): Add faster way to produce (#438)
When producing a large amount of messages, it can be that the profile
consists only of `Future()` calls, and most of the time is being spent
on running that ctor.

When you look inside of Future.__init__, all it does is create
threading.Condition so that `result()` can be blocking.

However, in the Produce step, we do not actually need synchronized
access to the future, nor do we need to have a blocking result(). We
only need done() and non-blocking result().

Introduce ab abstract interface for futures that is fulfilled by stdlib
Future() and a new class that does not offer performant result().