feat(producer): Add an AsyncProducer that returns Future (#476)
Create a new AsyncProducer that returns a Future on produce. Under the hood, the AsyncProducer is using the rdkafka::FutureProducer which uses channels to send the results from the delivery callback. This is also using the send_result function, which will error immediately if the message can't be put into the local rdkafka queue (e.g. if the queue is full).