ref(rust): Simplify metrics interface (breaking change)
The metrics interface in rust-arroyo was inherited from merni and has
support for value-less (or key-less) metric tags. Those are not used in
arroyo, so we can simplify the interface. We can also guarantee that
keys are not dynamically generated, saving us one trait object.

This will make it easier to interface with the metrics crate in the
streams repo, which does not support value-only tags.